renaming indexer service to celenium service, adding endpoints
