+++ celenium/service.go
@@ -1,10 +1,9 @@
-package indexer
+package celenium
 
 import (
 	"context"
 	"fmt"
 
-	"github.com/docker/compose/v2/pkg/api"
 	tc "github.com/testcontainers/testcontainers-go/modules/compose"
 	"go.uber.org/multierr"
 
@@ -16,7 +15,10 @@ import (
 )
 
 const (
-	IndexerServiceName = "indexer"
+	CeleniumServiceName = "celenium"
+	DatabaseLabel       = "celenium-database"
+	FrontendLabel       = "celenium-frontend"
+	BackendAPILabel     = "celenium-backend"
 )
 
 type Service struct {
@@ -37,7 +39,7 @@ func (s *Service) EndpointsNeeded() []string {
 }
 
 func (s *Service) EndpointsProvided() []string {
-	return []string{}
+	return []string{DatabaseLabel, FrontendLabel, BackendAPILabel}
 }
 
 func (s *Service) Setup(ctx context.Context, dir string, pendingGenesis *types.GenesisDoc) (genesis.Modifier, error) {
@@ -48,6 +50,8 @@ func (s *Service) Start(ctx context.Context, dir string, genesis *types.GenesisD
 	ctx, cancel := context.WithCancel(context.Background())
 	s.cancel = cancel
 
+	// TODO: this is a temporary solution to get the docker-compose file,
+	// but the command needs to be able to be run from anywhere on fs
 	compose, err := tc.NewDockerCompose("celenium/docker-compose.yml")
 	if err != nil {
 		return nil, fmt.Errorf("failed to create docker-compose: %w", err)
@@ -60,6 +64,12 @@ func (s *Service) Start(ctx context.Context, dir string, genesis *types.GenesisD
 		return nil, multierr.Combine(err, stopErr)
 	}
 
+	endpoints = apollo.Endpoints{
+		DatabaseLabel:   "http://127.0.0.1:5432",
+		FrontendLabel:   "http://127.0.0.1:3030",
+		BackendAPILabel: "http://127.0.0.1:9876",
+	}
+
 	return nil, err
 }
 
@@ -70,5 +80,5 @@ func (s *Service) Stop(ctx context.Context) error {
 }
 
 func (s *Service) Name() string {
-	return IndexerServiceName
+	return CeleniumServiceName
 }
